ボス敵を倒した時のエフェクトを追加
