+++ DirectXGame.vcxproj
@@ -62,7 +62,7 @@
       <ConformanceMode>true</ConformanceMode>
       <ProgramDataBaseFileName>$(IntDir)vc$(PlatformToolsetVersion).pdb</ProgramDataBaseFileName>
       <TreatWarningAsError>true</TreatWarningAsError>
-      <AdditionalIncludeDirectories>$(ProjectDir)DirectXGame\Engine\Sound\;$(ProjectDir)DirectXGame\Engine\Level\;$(ProjectDir)DirectXGame\Engine\Scene\;$(ProjectDir)DirectXGame\Application\Scene\;$(ProjectDir)DirectXGame\Engine\Input\;$(ProjectDir)DirectXGame\Engine\Math\;$(ProjectDir)DirectXGame\Application\GameObject\;$(ProjectDir)DirectXGame\Engine\Collision\;$(ProjectDir)DirectXGame\Engine\Camera\;$(ProjectDir)DirectXGame\Engine\Base\;$(ProjectDir)DirectXGame\Engine\3D\;$(ProjectDir)DirectXGame\Engine\2D\;$(ProjectDir)DirectXGame\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir)DirectXGame\Engine\Sound\;$(ProjectDir)DirectXGame\Engine\Level\;$(ProjectDir)DirectXGame\Engine\Scene\;$(ProjectDir)DirectXGame\Application\Scene\;$(ProjectDir)DirectXGame\Engine\Input\;$(ProjectDir)DirectXGame\Engine\Math\;$(ProjectDir)DirectXGame\Application\GameObject\Enemy\;$(ProjectDir)DirectXGame\Application\GameObject\Player\;$(ProjectDir)DirectXGame\Application\GameObject\;$(ProjectDir)DirectXGame\Engine\Collision\;$(ProjectDir)DirectXGame\Engine\Camera\;$(ProjectDir)DirectXGame\Engine\Base\;$(ProjectDir)DirectXGame\Engine\3D\;$(ProjectDir)DirectXGame\Engine\2D\;$(ProjectDir)DirectXGame\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <AdditionalOptions>/utf-8 %(AdditionalOptions)</AdditionalOptions>
       <LanguageStandard>stdcpp20</LanguageStandard>
     </ClCompile>
@@ -88,7 +88,7 @@
       <RuntimeLibrary>MultiThreaded</RuntimeLibrary>
       <ProgramDataBaseFileName>$(IntDir)vc$(PlatformToolsetVersion).pdb</ProgramDataBaseFileName>
       <TreatWarningAsError>true</TreatWarningAsError>
-      <AdditionalIncludeDirectories>$(ProjectDir)DirectXGame\Engine\Sound\;$(ProjectDir)DirectXGame\Engine\Level\;$(ProjectDir)DirectXGame\Engine\Scene\;$(ProjectDir)DirectXGame\Application\Scene\;$(ProjectDir)DirectXGame\Engine\Input\;$(ProjectDir)DirectXGame\Engine\Math\;$(ProjectDir)DirectXGame\Application\GameObject\;$(ProjectDir)DirectXGame\Engine\Collision\;$(ProjectDir)DirectXGame\Engine\Camera\;$(ProjectDir)DirectXGame\Engine\Base\;$(ProjectDir)DirectXGame\Engine\3D\;$(ProjectDir)DirectXGame\Engine\2D\;$(ProjectDir)DirectXGame\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir)DirectXGame\Engine\Sound\;$(ProjectDir)DirectXGame\Engine\Level\;$(ProjectDir)DirectXGame\Engine\Scene\;$(ProjectDir)DirectXGame\Application\Scene\;$(ProjectDir)DirectXGame\Engine\Input\;$(ProjectDir)DirectXGame\Engine\Math\;$(ProjectDir)DirectXGame\Application\GameObject\Enemy\;$(ProjectDir)DirectXGame\Application\GameObject\Player\;$(ProjectDir)DirectXGame\Application\GameObject\;$(ProjectDir)DirectXGame\Engine\Collision\;$(ProjectDir)DirectXGame\Engine\Camera\;$(ProjectDir)DirectXGame\Engine\Base\;$(ProjectDir)DirectXGame\Engine\3D\;$(ProjectDir)DirectXGame\Engine\2D\;$(ProjectDir)DirectXGame\;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <AdditionalOptions>/utf-8 %(AdditionalOptions)</AdditionalOptions>
       <LanguageStandard>stdcpp20</LanguageStandard>
     </ClCompile>
@@ -102,15 +102,16 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="DirectXGame\Application\GameObject\Enemy.cpp" />
-    <ClCompile Include="DirectXGame\Application\GameObject\EnemyBullet.cpp" />
-    <ClCompile Include="DirectXGame\Application\GameObject\InvEnemy.cpp" />
-    <ClCompile Include="DirectXGame\Application\GameObject\BossEnemy.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Enemy\BossEnemy.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Enemy\Enemy.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Enemy\EnemyBullet.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Enemy\InvEnemy.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Enemy\WeakEnemy.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Explosion.cpp" />
     <ClCompile Include="DirectXGame\Application\GameObject\Meteor.cpp" />
-    <ClCompile Include="DirectXGame\Application\GameObject\Player.cpp" />
-    <ClCompile Include="DirectXGame\Application\GameObject\PlayerBullet.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Player\Player.cpp" />
+    <ClCompile Include="DirectXGame\Application\GameObject\Player\PlayerBullet.cpp" />
     <ClCompile Include="DirectXGame\Application\GameObject\SkyDome.cpp" />
-    <ClCompile Include="DirectXGame\Application\GameObject\WeakEnemy.cpp" />
     <ClCompile Include="DirectXGame\Application\Scene\GameClearScene.cpp" />
     <ClCompile Include="DirectXGame\Application\Scene\GameOverScene.cpp" />
     <ClCompile Include="DirectXGame\Application\Scene\GamePlayScene.cpp" />
@@ -230,15 +231,16 @@
     </FxCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="DirectXGame\Application\GameObject\Enemy.h" />
-    <ClInclude Include="DirectXGame\Application\GameObject\EnemyBullet.h" />
-    <ClInclude Include="DirectXGame\Application\GameObject\InvEnemy.h" />
-    <ClInclude Include="DirectXGame\Application\GameObject\BossEnemy.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Enemy\BossEnemy.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Enemy\Enemy.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Enemy\EnemyBullet.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Enemy\InvEnemy.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Enemy\WeakEnemy.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Explosion.h" />
     <ClInclude Include="DirectXGame\Application\GameObject\Meteor.h" />
-    <ClInclude Include="DirectXGame\Application\GameObject\Player.h" />
-    <ClInclude Include="DirectXGame\Application\GameObject\PlayerBullet.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Player\Player.h" />
+    <ClInclude Include="DirectXGame\Application\GameObject\Player\PlayerBullet.h" />
     <ClInclude Include="DirectXGame\Application\GameObject\SkyDome.h" />
-    <ClInclude Include="DirectXGame\Application\GameObject\WeakEnemy.h" />
     <ClInclude Include="DirectXGame\Application\Scene\GameClearScene.h" />
     <ClInclude Include="DirectXGame\Application\Scene\GameOverScene.h" />
     <ClInclude Include="DirectXGame\Application\Scene\GamePlayScene.h" />
