+++ Resources/explosion01/explosion01.mtl
@@ -1,0 +1,10 @@
+newmtl マテリアル
+Ns 250.000000
+Ka 1.000000 1.000000 1.000000
+Kd 0.000000 0.000000 0.000000
+Ks 0.000000 0.000000 0.000000
+Ke 0.000000 0.000000 0.000000
+Ni 1.450000
+d 1.000000
+illum 2
+map_Kd texture.jpg
